feat: change Dijkstra weight from distance to time for shortest path calculation

+++ hub-service/src/main/java/com/winner/client/hubservice/hub/presentation/dto/HubNodeInfo.java
@@ -9,5 +9,5 @@ public record HubNodeInfo(
     String toHubName,
     int sequence,
     double estimatedDistance,
-    int estimatedArrivalTime
+    double estimatedArrivalTime
 ) {}
